# openapi/actionproxy.openapi.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "ActionProxy Community HTTP API",
    "version": "0.1.0",
    "description": "The self-hosted Community API for submitting governed tool calls, reviewing approvals, executing exact authorized actions, and verifying local audit evidence. Static web routes and routes outside the Community boundary are intentionally excluded.",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8787",
      "description": "Default loopback development server"
    }
  ],
...
  "paths": {
...
    "/v1/mcp/tool-calls": {
      "post": {
        "operationId": "postV1McpToolCalls",
        "summary": "Submit a tool call from the authenticated stdio MCP adapter",
        "tags": [
          "MCP adapter"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope or resource access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "description": "Required transport request identifier for the stdio MCP adapter.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Authenticated stdio MCP wrapper session UUID.",
            "in": "header",
            "name": "X-ActionProxy-MCP-Session-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitToolCall"
              }
            }
          }
        }
      }
    },
    "/v1/mcp/tool-calls/{id}": {
      "get": {
        "operationId": "getV1McpToolCallsId",
        "summary": "Get MCP-adapter tool-call status",
        "tags": [
          "MCP adapter"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCall"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope or resource access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "description": "Authenticated stdio MCP wrapper session UUID.",
            "in": "header",
            "name": "X-ActionProxy-MCP-Session-Id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/v1/tool-calls": {
      "get": {
        "operationId": "getV1ToolCalls",
        "summary": "List governed tool calls",
        "tags": [
          "Tool calls"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallList"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope or resource access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "query",
            "name": "decision",
            "required": false,
            "schema": {
              "enum": [
                "allow",
                "require_approval",
                "deny"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "maximum": 1000,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "runId",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "sessionId",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "enum": [
                "submitted",
                "authorized",
                "executed",
                "pending_approval",
                "blocked",
                "rejected",
                "failed"
              ],
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "toolName",
            "required": false,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "postV1ToolCalls",
        "summary": "Submit a governed tool call",
        "tags": [
          "Tool calls"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCallSubmission"
                }
              }
            }
          },
          "400": {
            "description": "Request validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope or resource access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "description": "Deduplicates a submission.",
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "maxLength": 512,
              "minLength": 1,
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitToolCall"
              }
            }
          }
        }
      }
    },
    "/v1/tool-calls/{id}": {
      "get": {
        "operationId": "getV1ToolCallsId",
        "summary": "Get a governed tool call",
        "tags": [
          "Tool calls"
        ],
        "security": [
          {
            "bearerAuth": []
          },
          {}
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ToolCall"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "The principal lacks the required scope or resource access",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "minLength": 1,
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "scheme": "bearer",
        "type": "http",
        "description": "A service-account API key or OIDC access token, depending on AUTH_MODE. Omit only when AUTH_MODE=none."
      }
    },
    "schemas": {
...
      "Error": {
        "properties": {
          "details": {
            "additionalProperties": true,
            "type": "object"
          },
          "error": {
            "minLength": 1,
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "ToolCallList": {
        "properties": {
          "toolCalls": {
            "items": {
              "$ref": "#/components/schemas/ToolCall"
            },
            "type": "array"
          }
        },
        "required": [
          "toolCalls"
        ],
        "type": "object"
      },
      "ToolCallSubmission": {
        "properties": {
          "approval": {
            "properties": {
              "id": {
                "minLength": 1,
                "type": "string"
              },
              "status": {
                "enum": [
                  "pending",
                  "approved",
                  "cancelled",
                  "expired",
                  "rejected"
                ],
                "type": "string"
              }
            },
            "required": [
              "id",
              "status"
            ],
            "type": "object"
          },
          "decision": {
            "enum": [
              "allow",
              "require_approval",
              "deny"
            ],
            "type": "string"
          },
          "error": {
            "type": "string"
          },
          "id": {
            "minLength": 1,
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "result": true,
          "risk": {
            "type": "string"
          },
          "status": {
            "enum": [
              "submitted",
              "authorized",
              "executed",
              "pending_approval",
              "blocked",
              "rejected",
              "failed"
            ],
            "type": "string"
          }
        },
        "required": [
          "decision",
          "id",
          "status"
        ],
        "type": "object"
      }
    }
  }
}